Update jquery.horizontalNav.js
Added: Option ```minimumItems``` to specify the minimum number of menu items required before setting the width. This should avoid having one or two items go super wide.
Also the class ```.horizontalNav-processed``` will be added to the supplied element if horizontalNav does is used, and ```.horizontalNav-notprocessed``` will be added if it's not. Useful for manually adding padding for when ```minimumItems``` is used. Also consider adding ```.horizontalNav-notprocessed``` to your raw html code to trigger padding in case of JS not loading.