Fix for fixed width menus
If I have to support dropdowns Firefox doesn't supports display:
table-cell and position: relative. So tableDisplay: false is needed.
Now, if the menu has fixed width (like in my case) all the menu items
are floating on eachother. With this fix the actual menu width can be
calculated even if the width is defined in CSS.